Set input

Creates a handler which can be used for onChange and similar props of 'React' components to set the value of a 'Shiny' input to one of the arguments passed to the handler.

setInput(inputId, argIdx = 1)

Arguments

inputId

'Shiny' input ID to set the value on.

argIdx

Index of the argument to use as value.

Value

A ReactData object which can be passed as a prop to 'React' components.